Refactored PackageTests to split out some common database testing functionality.
