+++ Tests/AppTests/Models/PackageTests.swift
@@ -4,26 +4,15 @@ import XCTest
 
 @testable import App
 
-final class PackageTests: XCTestCase
+final class PackageTests: DatabaseTestCase
 {
-  var app: Application!
-  var database: PostgreSQLConnection!
-
   override func setUp()
   {
-    try! Application.reset()
-    app = try! Application.testable()
-    database = try! app.newConnection(to: .psql).wait()
-
+    super .setUp()
+    
     // Create a couple of test packages that can be relied on during a test
     createTestPackage(from: "https://github.com/vapor/vapor.git")
     createTestPackage(from: "https://github.com/Alamofire/Alamofire.git")
-  }
-
-  override func tearDown()
-  {
-    database.close()
-    try? app.syncShutdownGracefully()
   }
 
   func testPackageCreationWithUrl() throws
@@ -43,15 +32,5 @@ final class PackageTests: XCTestCase
     XCTAssertThrowsError(try Package.findByUrl(on: database, url: nonExistentUrl).wait()) { error in
       XCTAssertEqual(error as? PackageError, PackageError.recordNotFound)
     }
-  }
-
-  /// Helper method to create test Package data.
-  /// - Parameter urlString: A string containing a valid URL.
-  func createTestPackage(from urlString: String)
-  {
-    guard let url = URL(string: urlString)
-      else { preconditionFailure("Expected a valid URL") }
-    let package = Package(url: url)
-    let _ = try! package.create(on: database).wait()
   }
 }
